chore(csproj): 🔧 remove redundant compile removal
