+++ SectigoCertificateManager/Clients/CertificatesClient.Core.cs
@@ -7,7 +7,6 @@ using System.IO;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Net.Http.Json;
-using System.Runtime.CompilerServices;
 using System.Text;
 using System.Text.Json;
 using System.Security.Cryptography.X509Certificates;
@@ -16,7 +15,7 @@ using SectigoCertificateManager.Utilities;
 /// <summary>
 /// Provides access to certificate related endpoints.
 /// </summary>
-public sealed class CertificatesClient {
+public sealed partial class CertificatesClient {
     private readonly ISectigoClient _client;
     private static readonly JsonSerializerOptions s_json = new() {
         PropertyNamingPolicy = JsonNamingPolicy.CamelCase
@@ -147,232 +146,6 @@ public sealed class CertificatesClient {
             .ReadFromJsonAsyncSafe<RenewCertificateResponse>(s_json, cancellationToken)
             .ConfigureAwait(false);
         return result?.SslId ?? 0;
-    }
-
-    /// <summary>
-    /// Searches for certificates using the provided filter.
-    /// </summary>
-    public async Task<CertificateResponse?> SearchAsync(
-        CertificateSearchRequest request,
-        CancellationToken cancellationToken = default) {
-        if (request is null) {
-            throw new ArgumentNullException(nameof(request));
-        }
-
-        var list = new List<Certificate>();
-        await foreach (var certificate in EnumerateSearchAsync(request, cancellationToken: cancellationToken).ConfigureAwait(false)) {
-            list.Add(certificate);
-        }
-
-        return list.Count == 0 ? null : new CertificateResponse { Certificates = list };
-    }
-
-    /// <summary>
-    /// Streams all certificates visible to the caller.
-    /// </summary>
-    /// <param name="pageSize">Number of certificates to request per page.</param>
-    /// <param name="cancellationToken">Token used to cancel the operation.</param>
-    public IAsyncEnumerable<Certificate> EnumerateCertificatesAsync(
-        int pageSize = 200,
-        CancellationToken cancellationToken = default) {
-        var request = new CertificateSearchRequest { Size = pageSize };
-        return EnumerateSearchAsync(request, cancellationToken);
-    }
-
-    /// <summary>
-    /// Streams issued certificates page by page and downloads each one.
-    /// </summary>
-    /// <param name="pageSize">Number of certificates to request per page.</param>
-    /// <param name="cancellationToken">Token used to cancel the operation.</param>
-    public IAsyncEnumerable<X509Certificate2> StreamCertificatesAsync(
-        int pageSize = 200,
-        CancellationToken cancellationToken = default) {
-        var request = new CertificateSearchRequest { Size = pageSize };
-        return StreamCertificatesAsync(request, cancellationToken);
-    }
-
-    /// <summary>
-    /// Streams issued certificates matching the provided filter.
-    /// </summary>
-    /// <param name="request">Filter describing certificates to retrieve.</param>
-    /// <param name="cancellationToken">Token used to cancel the operation.</param>
-    public async IAsyncEnumerable<X509Certificate2> StreamCertificatesAsync(
-        CertificateSearchRequest request,
-        [EnumeratorCancellation] CancellationToken cancellationToken = default) {
-        if (request is null) {
-            throw new ArgumentNullException(nameof(request));
-        }
-
-        await foreach (var certificate in EnumerateSearchAsync(request, cancellationToken).ConfigureAwait(false)) {
-            yield return await DownloadAsync(certificate.Id, cancellationToken).ConfigureAwait(false);
-        }
-    }
-
-    /// <summary>
-    /// Streams search results page by page.
-    /// </summary>
-    /// <param name="request">Filter describing certificates to retrieve.</param>
-    /// <param name="cancellationToken">Token used to cancel the operation.</param>
-    public async IAsyncEnumerable<Certificate> EnumerateSearchAsync(
-        CertificateSearchRequest request,
-        [EnumeratorCancellation] CancellationToken cancellationToken = default) {
-        if (request is null) {
-            throw new ArgumentNullException(nameof(request));
-        }
-
-        var originalSize = request.Size;
-        var originalPosition = request.Position;
-        var pageSize = request.Size ?? 200;
-        var position = request.Position ?? 0;
-
-        try {
-            var query = BuildQuery(request);
-            var response = await _client.GetAsync($"v1/certificate{query}", cancellationToken).ConfigureAwait(false);
-            var page = await response.Content
-                .ReadFromJsonAsyncSafe<IReadOnlyList<Certificate>>(s_json, cancellationToken)
-                .ConfigureAwait(false);
-            if (page is null || page.Count == 0) {
-                yield break;
-            }
-
-            foreach (var certificate in page) {
-                yield return certificate;
-            }
-
-            if (page.Count < pageSize) {
-                yield break;
-            }
-
-            request.Size = pageSize;
-            position += pageSize;
-
-            while (true) {
-                request.Position = position;
-                query = BuildQuery(request);
-                response = await _client.GetAsync($"v1/certificate{query}", cancellationToken).ConfigureAwait(false);
-                page = await response.Content
-                    .ReadFromJsonAsyncSafe<IReadOnlyList<Certificate>>(s_json, cancellationToken)
-                    .ConfigureAwait(false);
-                if (page is null || page.Count == 0) {
-                    yield break;
-                }
-
-                foreach (var certificate in page) {
-                    yield return certificate;
-                }
-
-                if (page.Count < pageSize) {
-                    yield break;
-                }
-
-                position += pageSize;
-            }
-        } finally {
-            request.Size = originalSize;
-            request.Position = originalPosition;
-        }
-    }
-
-    /// <summary>
-    /// Downloads an issued certificate and returns an <see cref="X509Certificate2"/> instance.
-    /// </summary>
-    /// <param name="certificateId">Identifier of the certificate to download.</param>
-    /// <param name="cancellationToken">Token used to cancel the operation.</param>
-    public async Task<X509Certificate2> DownloadAsync(
-        int certificateId,
-        CancellationToken cancellationToken = default) {
-        if (certificateId <= 0) {
-            throw new ArgumentOutOfRangeException(nameof(certificateId));
-        }
-
-        var endpoint = $"ssl/v1/collect/{certificateId}";
-        var url = $"{endpoint}?format=base64";
-        var response = await _client.GetAsync(url, cancellationToken).ConfigureAwait(false);
-#if NETSTANDARD2_0 || NET472
-        var base64 = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
-#else
-        var base64 = await response.Content.ReadAsStringAsync(cancellationToken).ConfigureAwait(false);
-#endif
-        return Certificate.FromBase64(base64);
-    }
-
-    /// <summary>
-    /// Downloads an issued certificate and saves it to disk.
-    /// </summary>
-    /// <param name="certificateId">Identifier of the certificate to download.</param>
-    /// <param name="path">Destination file path.</param>
-    /// <param name="format">Certificate format to request. Defaults to <c>base64</c>.</param>
-    /// <param name="progress">Optional progress reporter.</param>
-    /// <param name="cancellationToken">Token used to cancel the operation.</param>
-    public async Task DownloadAsync(
-        int certificateId,
-        string path,
-        string format = "base64",
-        IProgress<double>? progress = null,
-        CancellationToken cancellationToken = default) {
-        if (certificateId <= 0) {
-            throw new ArgumentOutOfRangeException(nameof(certificateId));
-        }
-        if (string.IsNullOrEmpty(path)) {
-            throw new ArgumentException("Path cannot be null or empty.", nameof(path));
-        }
-
-        var endpoint = $"ssl/v1/collect/{certificateId}";
-        var url = $"{endpoint}?format={Uri.EscapeDataString(format)}";
-        var response = await _client.GetAsync(url, cancellationToken).ConfigureAwait(false);
-        var stream = await response.Content.ReadAsStreamAsync().ConfigureAwait(false);
-        using (stream) {
-            Directory.CreateDirectory(Path.GetDirectoryName(path)!);
-            using var file = new FileStream(path, FileMode.Create, FileAccess.Write, FileShare.None);
-            var buffer = new byte[81920];
-            long total = response.Content.Headers.ContentLength ?? (stream.CanSeek ? stream.Length : -1);
-            long copied = 0;
-            int count;
-            while (true) {
-#if NETSTANDARD2_0 || NET472
-                count = await stream.ReadAsync(buffer, 0, buffer.Length).ConfigureAwait(false);
-#else
-                count = await stream.ReadAsync(buffer.AsMemory(0, buffer.Length), cancellationToken).ConfigureAwait(false);
-#endif
-                if (count == 0) {
-                    break;
-                }
-#if NETSTANDARD2_0 || NET472
-                await file.WriteAsync(buffer, 0, count).ConfigureAwait(false);
-#else
-                await file.WriteAsync(buffer.AsMemory(0, count), cancellationToken).ConfigureAwait(false);
-#endif
-                copied += count;
-                if (progress is not null && total > 0) {
-                    progress.Report((double)copied / total);
-                }
-            }
-
-            if (progress is not null && total > 0) {
-                progress.Report(1d);
-            }
-        }
-    }
-
-    /// <summary>
-    /// Downloads the issuing certificate chain and saves it to disk.
-    /// </summary>
-    /// <param name="certificateId">Identifier of the certificate.</param>
-    /// <param name="path">Destination file path.</param>
-    /// <param name="cancellationToken">Token used to cancel the operation.</param>
-    public async Task GetCaChainAsync(
-        int certificateId,
-        string path,
-        CancellationToken cancellationToken = default) {
-        if (certificateId <= 0) {
-            throw new ArgumentOutOfRangeException(nameof(certificateId));
-        }
-        if (string.IsNullOrEmpty(path)) {
-            throw new ArgumentException("Path cannot be null or empty.", nameof(path));
-        }
-
-        await DownloadAsync(certificateId, path, format: "x509IO", cancellationToken: cancellationToken)
-            .ConfigureAwait(false);
     }
 
     /// <summary>
