+++ SectigoCertificateManager/SectigoCertificateManager.csproj
@@ -17,4 +17,5 @@
     <PackageReference Include="System.Formats.Asn1" Version="9.0.7" />
   </ItemGroup>
 
+
 </Project>
